Ocean module

See merge request public_projects/kadlu!43

+++ kadlu/geospatial/data_sources/chs.py
@@ -1,25 +1,27 @@
-""" 
-    Module within the kadlu package for handling Non-Navigational NONNA-100 
-    bathymetric data from the The Canadian Hydrographic Service (CHS). 
-
-    Authors: Oliver Kirsebom, Casey Hilliard, Matthew Smith
-    contact: oliver.kirsebom@dal.ca
-    Organization: MERIDIAN-Intitute for Big Data Analytics
-    Team: Acoustic data Analytics, Dalhousie University
-    Project: packages/kadlu
-             Project goal: Tools for underwater soundscape modeling
+"""
+    API for NONNA-100 Bathymetric Data from the Canadian Hydrographic Service (CHS)
+     
+    Metadata regarding the dataset can be found here:
+        https://open.canada.ca/data/en/dataset/d3881c4c-650d-4070-bf9b-1e00aabf0a1d
 """
 
 import os
-import numpy as np
-from kadlu.geospatial.geospatial import crop, read_geotiff
 import json
 import requests
-from osgeo import gdal
 import warnings
+from datetime import datetime
 
-from kadlu.geospatial.data_sources.fetch_util import \
-storage_cfg, database_cfg, chs_table, str_def
+import numpy as np
+from osgeo import gdal
+
+import kadlu.geospatial.data_sources.fetch_handler
+from kadlu.geospatial.data_sources.data_util        import          \
+        database_cfg,                                               \
+        storage_cfg,                                                \
+        insert_hash,                                                \
+        serialized,                                                 \
+        chs_table,                                                  \
+        str_def
 
 
 conn, db = database_cfg()
@@ -35,28 +37,19 @@ def parse_sw_corner(path):
     return south, west
 
 
-# matt_s 2019-12
-# removed this function since filenames are now pulled directly from source
-
-#def filename(south, west):
-#    """ generate a filename for given southwest corner coordinates """
-#    fname = "CA2_{0:04d}N{1:05d}W.tif".format(int(south * 100), -int(west * 100))
-#    return fname
-
-
 def fetch_chs(south, north, west, east, band_id=1):
     """ download bathymetric geotiffs, process them, and insert into db
 
-    args:
-        south, north: float
-            ymin, ymax coordinate boundaries to fetch bathymetry. range: -90, 90
-        west, east: float
-            xmin, xmax coordinate boundaries to fetch bathymetry. range: -180, 180
+        args:
+            south, north: float
+                ymin, ymax coordinate boundaries. range: -90, 90
+            west, east: float
+                xmin, xmax coordinate boundaries. range: -180, 180
 
-    return: 
-      nothing
-
+        return: 
+            True if new data was downloaded and processed, else False
     """
+
     # api call: get raster IDs within bounding box
     source = "https://gisp.dfo-mpo.gc.ca/arcgis/rest/services/FGP/CHS_NONNA_100/"
     spatialRel = "esriSpatialRelIntersects"
@@ -83,22 +76,25 @@ def fetch_chs(south, north, west, east, band_id=1):
     # api call: for each tiff image, download the associated rasters
     filepaths = []
     imgnum = 1
-    print()
     for img in imgs:
         fname = img['id'].split('\\')[-1]
         fpath = f"{storage_cfg()}{fname}"
-        print(f"{fname}: downloading {imgnum}/{len(imgs)} from CHS NONNA-100...", end="\r")
-        #if os.path.isfile(fpath): continue
         filepaths.append(fpath)
-        assert(len(img['rasterIds']) == 1)
-        url3 = f"{source}ImageServer/file?id={img['id'][0:]}&rasterId={img['rasterIds'][0]}"
-        tiff = requests.get(url3)
-        assert(tiff.status_code == 200)
-        with open(fpath, "wb") as f: f.write(tiff.content)
-        imgnum += 1
-        
-    print()
-    assert(len(filepaths) > 0)
+        if os.path.isfile(fpath): 
+            #print(f'CHS {fname} bathymetry: file found, skipping download')
+            pass
+        else:
+            print(f"CHS {fname} bathymetry: downloading {imgnum}/{len(imgs)} "
+                   "from CHS NONNA-100...")
+            assert(len(img['rasterIds']) == 1)
+            url3 = f"{source}ImageServer/file?id={img['id'][0:]}&rasterId={img['rasterIds'][0]}"
+            tiff = requests.get(url3)
+            assert(tiff.status_code == 200)
+            with open(fpath, "wb") as f: f.write(tiff.content)
+            imgnum += 1
+
+    print(f'CHS bathymetry: processing {len(filepaths)} '
+          f'file{"s" if len(filepaths)!=1 else ""}')
 
     # read downloaded files and process them for DB insertion
     for filepath in filepaths:
@@ -121,31 +117,56 @@ def fetch_chs(south, north, west, east, band_id=1):
         file_lat = np.linspace(start=file_south, stop=file_ymax, num=tiff_data.RasterYSize)
         file_lon = np.linspace(start=file_west,  stop=file_xmax, num=tiff_data.RasterXSize)
 
-        # select non-masked entries and remove missing
+        # select non-masked entries, remove missing, build grid
         z1 = np.flip(bathy, axis=0)
         x1, y1 = np.meshgrid(file_lon, file_lat)
         ix = z1[~z1.mask] != band.GetNoDataValue()
         x2 = x1[~z1.mask][ix]
         y2 = y1[~z1.mask][ix]
-        z2 = np.array(z1[~z1.mask][ix].data, dtype=int)
-
-        # build coordinate grid and insert into db
+        z2 = np.abs(z1[~z1.mask][ix].data)
         source = ['chs' for z in z2]
         grid = list(map(tuple, np.vstack((z2, y2, x2, source)).T))
+
+        # insert into db
         n1 = db.execute(f"SELECT COUNT(*) FROM {chs_table}").fetchall()[0][0]
         db.executemany(f"INSERT OR IGNORE INTO {chs_table} VALUES (?,?,?,?)", grid)
         n2 = db.execute(f"SELECT COUNT(*) FROM {chs_table}").fetchall()[0][0]
         db.execute("COMMIT")
         conn.commit()
-
-        print(f"{filepath.split('/')[-1]} processed and inserted {n2-n1} rows. "
+        print(f"CHS {filepath.split('/')[-1]} bathymetry: "
+              f"processed and inserted {n2-n1} rows. "
               f"{len(z1[~z1.mask]) - len(grid)} null values removed, "
               f"{len(grid) - (n2-n1)} duplicate rows ignored")
 
-    return 
+    return True
 
 
 def load_chs(south, north, west, east):
+    """ load bathymetric data from the database
+
+        args:
+            south, north:
+                y-grid coordinate boundaries (float)
+            west, east:
+                x-grid coordinate boundaries (float)
+
+        return:
+            bathy:
+                bathymetric values within query range
+            lat:
+                y-grid coordinate values
+            lon:
+                x-grid coordinate values
+    """
+    # check for missing data
+    qryargs = dict(
+            south=south, west=west,
+            north=north, east=east, 
+            start=datetime.now(), end=datetime.now())
+    kadlu.geospatial.data_sources.fetch_handler.fetch_handler(
+            'bathy', 'chs', parallel=1, **qryargs)
+
+    # load the data
     db.execute(' AND '.join([f"SELECT * FROM {chs_table} WHERE lat >= ?",
                                                               "lat <= ?",
                                                               "lon >= ?",
@@ -153,36 +174,35 @@ def load_chs(south, north, west, east):
                tuple(map(str, [south, north, west, east])))
     
     slices = np.array(db.fetchall(), dtype=object).T
-    #assert len(slices) == 4, "no data found for query range"
-    if len(slices) != 4:
-       warnings.warn("no data found for query range, returning empty arrays")
-       bathy, lat, lon = np.array([]), np.array([]), np.array([])
-    else:
-       bathy, lat, lon, source = slices
-    return bathy, lat, lon
+    assert len(slices) == 4, "no data found for query range"
+    bathy, lat, lon, source = slices
+    return np.array((bathy, lat, lon)).astype(float)
 
 
 class Chs():
-    """ collection of module functions for fetching and loading. abstracted to include a seperate function for each variable """
+    """ collection of module functions for fetching and loading """
 
-    """
-    def fetch_bathymetry(self, south=44.4, north=44.7, west=-64.4, east=-63.8):
-        return fetch_chs(south, north, west, east, band_id=1)
-
-    def load_bathymetry(self, south=44.4, north=44.7, west=-64.4, east=-63.8):
-        return load_chs(south, north, west, east)
-    """
     def fetch_bathymetry(self, **kwargs):
-        return fetch_chs(south=kwargs['south'], north=kwargs['north'], 
-              west=kwargs['west'], east=kwargs['east'], band_id=1)
+        # trim query indexing entropy and check for fetched data
+        for k in ('start', 'lock', 'end', 'top', 'bottom'):
+            if k in kwargs.keys(): del kwargs[k]
+        if serialized(kwargs, 'fetch_chs_bathy'): return False
+
+        # if new data was fetched, index the query hash
+        if (fetch_chs(south=kwargs['south'], north=kwargs['north'], 
+                west=kwargs['west'], east=kwargs['east'], band_id=1)):
+            insert_hash(kwargs, 'fetch_chs_bathy')
+        return True
 
     def load_bathymetry(self, **kwargs):
         return load_chs(south=kwargs['south'], north=kwargs['north'], 
               west=kwargs['west'], east=kwargs['east'])
 
     def __str__(self):
-        info = "Non-Navigational 100m (NONNA-100) bathymetry dataset from Canadian Hydrographic Datastore"
-        args = "(south=-90, north=90, west=-180, east=180)"
+        info = "\n".join(["Non-Navigational 100m (NONNA-100) bathymetry dataset",
+            "from the Canadian Hydrographic Datastore",
+            "\thttps://open.canada.ca/data/en/dataset/d3881c4c-650d-4070-bf9b-1e00aabf0a1d"])
+        args = "(south, north, west, east)"
         return str_def(self, info, args)
